Prevent duplicate files in Destination project
fixes #3.
If a file exists the destination project then this will skip linking to a file in the source project that has the same path. Handy for overwriting things in the destination. Also keeps a running tally and will not re-link a file with the same path, which shouldn't really happen anyway but whoa.